Rename domain -> bloc and presentation -> ui
