+++ lib/feature/auth/signup/ui/signup_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:flutter_template/feature/auth/signup/domain/signup.dart';
+import 'package:flutter_template/feature/auth/signup/bloc/signup.dart';
 import 'package:flutter_template/widgets/circular_progress_indicator.dart';
 
 class SignUpPage extends StatelessWidget {
